Updated CSS for Eventmanage & Event pages
I updated the css files for events since the announcements at the bottom of the event page were to wide and not as appealing so i made them thinner. The event management page i edited both tables as the contents were too long and not aligned properly in the table nor the screen.